+++ src/main/resources/static/err.html
@@ -5,6 +5,6 @@
     <title>Err</title>
 </head>
 <body>
-<h1>Błąd przy dodawaniu użytkownika!</h1>
+<h1>Something went wrong! Check parameters in URL</h1>
 </body>
 </html>
