+++ src/main/resources/static/success.html
@@ -5,6 +5,6 @@
     <title>Success</title>
 </head>
 <body>
-<h1>Dodano użytkownika!</h1>
+<h1>User has been added successfully!</h1>
 </body>
 </html>
